Add Review Voting, Fix TABLE Review (int vote -> Float vote)

+++ unicat_online_learning/src/main/java/com/unicat/onlinelearning/dto/Review.java
@@ -13,12 +13,12 @@ public class Review {
     private int ReviewID;
     private int UserID;
     private int CourseID;
-    private int Vote;
+    private float Vote;
 
     public Review() {
     }
 
-    public Review(int ReviewID, int UserID, int CourseID, int Vote) {
+    public Review(int ReviewID, int UserID, int CourseID, float Vote) {
         this.ReviewID = ReviewID;
         this.UserID = UserID;
         this.CourseID = CourseID;
@@ -49,11 +49,11 @@ public class Review {
         this.CourseID = CourseID;
     }
 
-    public int getVote() {
+    public float getVote() {
         return Vote;
     }
 
-    public void setVote(int Vote) {
+    public void setVote(float Vote) {
         this.Vote = Vote;
     }
